所有Entity类修改完成，编译和测试均通过。 修改汇总 修改类型	数量	说明 主键重命名	3个	UserRole/RolePerm/UserDepartment 的 xxx_id → id 新增 is_deleted	36个	所有业务表Entity 新增 version	42个	所有业务表Entity 新增业务字段	10个	Department/Unit/DosageForm/PurchaseOrders/PurchaseSuppliers/StockIn/StockOut/ProdUnitMaterialFile Controller修复	3个	PurchaseOrdersController/PurchaseSuppliersController/RolePermController Service修复	2个	PurchaseSuppliersService/PurchaseSuppliersServiceImpl

+++ src/main/java/com/tonghui/erp/Controller/PurchaseOrdersController.java
@@ -56,7 +56,7 @@ public class PurchaseOrdersController extends BaseCrudController<PurchaseOrders,
 
     @Override
     protected PurchaseOrders doUpdate(Long id, PurchaseOrders purchaseOrders) {
-        purchaseOrders.setId(id.intValue()); // 转换为Integer类型
+        purchaseOrders.setId(id);
         purchaseOrdersService.updatePurchaseOrder(purchaseOrders);
         return purchaseOrders;
     }
